==> 00_NAKO/01_Load_data/00_load_demographics_and_merge_label.ipynb <==
import pandas as pd
import os

csv_path = "/zi/flstorage/HITKIP/datasets/NAKO_01_2024/NAKO/sourcedata/NAKO-711_Studiendaten/NAKO-711_export_Auswertung_export_baseline.csv"
merged_df = pd.read_csv(csv_path, sep=";")
merged_df["Id"] = merged_df["ID"].astype(str)

merged_df.to_csv("../00_data/NAKO_all_data.csv", index=False)

mri_data = "/zi/flstorage/HITKIP/datasets/NAKO_01_2024/NAKO/derivatives/ses-01/freesurfer_stats/"
#load specific files to ../00_data/mris if they contain 
output_dir = "../00_data/mris/"
os.makedirs(output_dir, exist_ok=True)
files = ["aparc.volume", "aparc.thickness", "aseg.volume"]
for file in files:
    for root, dirs, files in os.walk(mri_data):
        for name in files:
            if file in name:
                print(name)
                df = pd.read_csv(os.path.join(root, name), sep="\t")
                #chang ending to csv
                name = name.replace(".tsv", ".csv")
                df.to_csv(os.path.join(output_dir, name), index=False)

demographics_covariates = merged_df[["ID", "basis_sex", "basis_age", "basis_uort"]]
#subtract 1 from sex
demographics_covariates["basis_sex"] = demographics_covariates["basis_sex"] - 1

#rename demographics_covariates to demographics_covariates
demographics_covariates.rename(
    columns={"ID": "ID", "basis_sex": "Sex", "basis_uort": "Site", "basis_age": "Age"},
    inplace=True,
)

demographics_covariates.to_csv("../00_data/age_label/all_ages_all_ids.csv", index=False)

demographics_covariates.to_csv("../00_data/NAKO_demographics_covariates.csv", index=False)

demographics_covariates
