==> 00_NAKO/01_Load_data/01_subset_age.ipynb <==
import pandas as pd

df = pd.read_csv("/zi/home/esra.lenz/Documents/00_HITKIP/09_TABPFN/00_NAKO/00_data/NAKO_demographics_covariates.csv")

print(df["Age"].describe())


df["Age"].hist()

df["Sex"].value_counts()

#box blot
df.boxplot(column="Age")

#just people from 18 to 26
df_young = df[(df["Age"] >= 18) & (df["Age"] <= 26)]
len(df_young)
# didnt do that yet Could be an experiment
